+++ src/migrations/20250610124829-replenishment_link.ts
@@ -21,19 +21,30 @@ export class Ureplenishment_link20250610124829 implements MigrationInterface {
             true,
         );
 
-        await queryRunner.createForeignKey(
-            'replenishment_link',
-            new TableForeignKey({
-                columnNames: ['work_reference'],
-                referencedColumnNames: ['work_reference'],
-                referencedTableName: 'pick_work_header',
-                onDelete: 'CASCADE',
-            }),
-        );
+        const pickWorkHeaderTableExists = await queryRunner.hasTable('pick_work_header');
+        if (pickWorkHeaderTableExists) {
+            await queryRunner.createForeignKey(
+                'replenishment_link',
+                new TableForeignKey({
+                    columnNames: ['work_reference'],
+                    referencedColumnNames: ['work_reference'],
+                    referencedTableName: 'pick_work_header',
+                    onDelete: 'CASCADE',
+                }),
+            );
+        } else {
+            console.log('Skipping foreign key creation for work_reference as the pick_work_header table does not exist yet.');
+        }
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.dropForeignKey('replenishment_link', 'FK_replenishment_link_work_reference');
-        await queryRunner.dropTable('replenishment_link');
+        const table = await queryRunner.getTable('replenishment_link');
+        if (table) {
+            const foreignKeys = table.foreignKeys;
+            for (const foreignKey of foreignKeys) {
+                await queryRunner.dropForeignKey('replenishment_link', foreignKey);
+            }
+            await queryRunner.dropTable('replenishment_link');
+        }
     }
 }
